Bring back the client's context pointer
This was removed, presumably due to lack of use. The caller can
presumably just have a global variable for its context so long as
there is only one, which there should be, except perhaps when
testing.

However, this is a bit ugly. It seems better to provide the context
to avoid the need for a global. Add it back.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I9a828f673ad95aabdc32988445e24e0f9774c221