Avoid using malloc() directly
This library is supposed to stand alone, so has avoided making use of
standard library functions for the most part. Recent development has
resulted in malloc() being used, presumably in cases where there is
no chance of a memory leak.

However the bootloader may wish to implement this differently from the
C-library version, or track the memory usage of vboot. Follow the
existing approach of using a vb2ex...() function for this, to permit
this flexibility.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I1c330f7e7ada7e6f4d0ec0da97dba3658aff460a