Allow use of fallthrough directly
There is a standard way of handling fallthrough, at least in linux and
U-Boot, of defining 'fallthrough' as the required attribute. At
present, vboot always defines this itself, which is incompatible with
this convention.

Arguably vboot should not mess with this at all. But for now, just
make it optional.

BUG=nona
BRANCH=none
TEST=none

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Id62f126f77311360fb7859ee956f1476378d87c0